Respond to PR comments.
