+++ tests/test_index.py
@@ -190,9 +190,8 @@ class TestIndex:
             _ = t.order_by(t.img.similarity(t.split)).limit(1).collect()
         assert 'not an expression' in str(exc_info.value).lower()
 
-        with pytest.raises(pxt.Error) as exc_info:
+        with pytest.raises(pxt.Error, match='No indices found for '):
             _ = t.order_by(t.split.similarity('red truck')).limit(1).collect()
-        assert 'no index found' in str(exc_info.value).lower()
 
         t = small_img_tbl
         t.add_embedding_index('img', image_embed=clip_embed)
@@ -204,6 +203,15 @@ class TestIndex:
         with pytest.raises(pxt.Error) as exc_info:
             _ = t.order_by(t.img.similarity('red truck')).limit(1).collect()
         assert "column 'img' has multiple indices" in str(exc_info.value).lower()
+
+        # Similarity fails when attempted on a snapshot
+        t_s = pxt.create_snapshot('t_s', t)
+        with pytest.raises(pxt.Error, match='No indices found for '):
+            _ = t_s.order_by(t_s.img.similarity('red truck')).limit(1).collect()
+
+        # Direct access to the unnamed embedding column fails on a snapshot
+        with pytest.raises(pxt.Error, match='No indices found for '):
+            r = t_s.select(t_s.img.embedding(idx='other_idx')).limit(2)
 
         t.drop_embedding_index(idx_name='idx0')
         t.drop_embedding_index(idx_name='idx1')
@@ -381,6 +389,11 @@ class TestIndex:
         res = img_t.where(img_t.img == rows[0]['img']).collect()
         assert len(res) == 1
 
+        # Direct access to the unnamed embedding column works
+        res = img_t.select(img_t.img.embedding()).limit(2).collect()
+        assert len(res) == 2
+        assert isinstance(res[0, 'img_embedding_'], np.ndarray)
+
         with pytest.raises(pxt.Error) as exc_info:
             # duplicate name
             img_t.add_embedding_index('img', idx_name='idx0', image_embed=clip_embed)
@@ -445,6 +458,19 @@ class TestIndex:
         sim = img_t.img.similarity('red truck', idx='other_idx')
         res = img_t.order_by(sim, asc=False).limit(1).collect()
         assert len(res) == 1
+
+        # Direct access to the new named embedding column works
+        r = img_t.select(img_t.img.embedding(idx='other_idx')).limit(2).collect()
+        assert len(r) == 2
+        assert isinstance(r[0, 'img_embedding_other_idx'], np.ndarray)
+
+        # Direct access to an unnamed embedding column fails when multiple indices are present
+        with pytest.raises(pxt.Error, match='has multiple indices'):
+            _ = img_t.select(img_t.img.embedding()).collect()
+
+        # Adding an index with an invalid index name fails
+        with pytest.raises(pxt.Error, match='Invalid column name'):
+            img_t.add_embedding_index(img_t.img, idx_name='BOGUS COL NAME', embedding=clip_embed)
 
         with pytest.raises(pxt.Error) as exc_info:
             _ = img_t.img.similarity('red truck', idx='doesnotexist')
@@ -514,54 +540,11 @@ class TestIndex:
         # the table in the ColumnRef, which is the base table.
         # So it raises error that there's no index.
         # Fix needs discussion.
-        with pytest.raises(pxt.Error) as exc_info:
+        with pytest.raises(pxt.Error, match='No indices found for '):
             df = v.select(sim=v.s.similarity(sents[1]))
-        assert 'no index found for column' in str(exc_info.value).lower()
         _ = reload_tester.run_query(v.select())
 
         reload_tester.run_reload_test()
-
-    def test_embedding_access(
-        self,
-        img_tbl: pxt.Table,
-        test_tbl: pxt.Table,
-        clip_embed: func.Function,
-        e5_embed: func.Function,
-        all_mpnet_embed: func.Function,
-        reload_tester: ReloadTester,
-    ) -> None:
-        skip_test_if_not_installed('transformers')
-        img_t = img_tbl
-        rows = list(img_t.select(img=img_t.img.fileurl, category=img_t.category, split=img_t.split).collect())
-        # create table with fewer rows to speed up testing
-        schema = {'img': pxt.Image, 'category': pxt.String, 'split': pxt.String}
-        tbl_name = 'index_test'
-        img_t = pxt.create_table(tbl_name, schema)
-        img_t.insert(rows[:30])
-        dummy_img_t = pxt.create_table('dummy', schema)
-        dummy_img_t.insert(rows[:10])
-
-        img_t.add_embedding_index('img', embedding=clip_embed)
-
-        r = img_t.select(img_t.img.embedding()).limit(2).collect()
-        assert len(r) == 2
-        assert isinstance(r[0, 'img_embedding_'], np.ndarray)
-
-        # multiple indices
-        img_t.add_embedding_index(img_t.img, idx_name='other_idx', embedding=clip_embed)
-        r = img_t.select(img_t.img.embedding('other_idx')).limit(2).collect()
-        assert len(r) == 2
-        assert isinstance(r[0, 'img_embedding_other_idx'], np.ndarray)
-
-        r = img_t.select(img_t.img.embedding(idx='other_idx')).limit(2).collect()
-        assert len(r) == 2
-        assert isinstance(r[0, 'img_embedding_other_idx'], np.ndarray)
-
-        with pytest.raises(pxt.Error, match='has multiple indices'):
-            _ = img_t.select(img_t.img.embedding()).collect()
-
-        with pytest.raises(pxt.Error, match='Invalid column name'):
-            img_t.add_embedding_index(img_t.img, idx_name='BOGUS COL NAME', embedding=clip_embed)
 
     def test_embedding_errors(self, small_img_tbl: pxt.Table, test_tbl: pxt.Table, clip_embed: func.Function) -> None:
         skip_test_if_not_installed('transformers')
